chore: reduce number of api calls in editor page
