+++ src/utils/EditorUtils.ts
@@ -1,6 +1,9 @@
 import { INITIAL_CONTEXT_DATA } from 'constants/common';
 import { CUSTOM_NODES } from 'constants/editor';
+import { toPng } from 'html-to-image';
+import { Snippet, SnippetData } from 'interfaces/AppProvider.interface';
 import { NodeDataStore } from 'interfaces/Editor.interface';
+import { cloneDeep, isEqual } from 'lodash';
 import { v4 as generateUniqueId } from 'uuid';
 
 export const getDragHandleByNodeType = (type: string) => {
@@ -50,4 +53,41 @@ export const getStringifiedValue = (data: unknown) => {
 
 export const getParsedValue = (data: string) => {
   return JSON.parse(data);
+};
+
+export const getSnippetSnapshot = async () => {
+  const node = document.querySelector('.react-flow');
+  if (node) {
+    try {
+      const imageBase64Url = await toPng(node as HTMLElement, {
+        filter: (node) => {
+          // we don't want to add the minimap and the controls to the image
+          if (
+            node?.classList?.contains('react-flow__minimap') ||
+            node?.classList?.contains('react-flow__controls')
+          ) {
+            return false;
+          }
+
+          return true;
+        },
+        quality: 1,
+      });
+
+      return imageBase64Url;
+    } catch (error) {
+      return '';
+    }
+  }
+
+  return '';
+};
+
+export const isAllowedToUpdate = (
+  snippetData: SnippetData,
+  updatedData: Partial<Snippet>
+) => {
+  const clonedSnippetData = cloneDeep(snippetData);
+  const updatedSnippetData = { ...clonedSnippetData, ...updatedData };
+  return !isEqual(clonedSnippetData, updatedSnippetData);
 };
